+++ components/auth/AuthFormError.tsx
@@ -1,14 +1,17 @@
 'use client'
 
 import type { ReactNode } from 'react'
+import { CircleAlert } from 'lucide-react'
 
 /**
- * Inline error alert for the auth forms (login, register, reset).
+ * Inline error line for the auth forms (login, register, reset).
  *
- * Auth failures render here, adjacent to the fields, instead of in a toast:
+ * Auth failures render here, adjacent to the form, instead of in a toast:
  * a toast in the corner auto-dismisses, sits far from the locus of attention,
- * and is easy to miss entirely. role="alert" makes screen readers announce
- * the message when it appears.
+ * and is easy to miss entirely. Styled as one quiet destructive sentence with
+ * an icon, mirroring the AttnLine pattern, never as a boxed banner: the box
+ * reads louder than the message and breaks the panel's rhythm. role="alert"
+ * makes screen readers announce the message when it appears.
  */
 export function AuthFormError({
   message,
@@ -18,14 +21,15 @@ export function AuthFormError({
   action?: ReactNode
 }) {
   return (
-    <div
+    <p
       role="alert"
-      className="animate-fade-in rounded-lg border border-destructive/30 bg-destructive/5 p-4"
+      className="animate-fade-in flex items-start gap-2 text-[13px] leading-5 text-destructive"
     >
-      <p className="text-sm text-destructive">
+      <CircleAlert className="mt-0.5 h-3.5 w-3.5 shrink-0" aria-hidden="true" />
+      <span>
         {message}
         {action && <> {action}</>}
-      </p>
-    </div>
+      </span>
+    </p>
   )
 }
