+++ messages/en.json
@@ -268,6 +268,10 @@
     "no_account": "Create account",
     "or_divider": "or",
     "or_email_divider": "or sign in with email",
+    "login_new_here": "New here?",
+    "method_email_chip": "Email",
+    "show_password": "Show password",
+    "hide_password": "Hide password",
     "login_failed_title": "Sign in failed",
     "login_failed_bankid": "Could not complete BankID sign in.",
     "login_invalid_credentials": "Wrong email address or password.",
