+++ messages/sv.json
@@ -268,6 +268,10 @@
     "no_account": "Skapa konto",
     "or_divider": "eller",
     "or_email_divider": "eller logga in med e-post",
+    "login_new_here": "Ny här?",
+    "method_email_chip": "E-post",
+    "show_password": "Visa lösenord",
+    "hide_password": "Dölj lösenord",
     "login_failed_title": "Inloggning misslyckades",
     "login_failed_bankid": "Kunde inte slutföra BankID-inloggningen.",
     "login_invalid_credentials": "Fel e-postadress eller lösenord.",
